+++ dflib-parquet/src/main/java/org/dflib/parquet/read/converter/StoringConverter.java
@@ -36,12 +36,12 @@ public interface StoringConverter {
         return (ValueHolder<?>) store();
     }
 
-    static StoringConverter of(Type colSchema, boolean accum) {
+    static StoringConverter of(Type colSchema, boolean accum, boolean dictionarySupport) {
 
         boolean allowsNulls = colSchema.getRepetition() == Type.Repetition.OPTIONAL;
 
         if (colSchema.isPrimitive()) {
-            return primitiveConverter(colSchema.asPrimitiveType(), accum, allowsNulls);
+            return primitiveConverter(colSchema.asPrimitiveType(), accum, dictionarySupport, allowsNulls);
         } else if (colSchema instanceof GroupType gt) {
             return groupConverter(gt, accum, allowsNulls);
         }
@@ -49,7 +49,11 @@ public interface StoringConverter {
         throw new RuntimeException(colSchema + " deserialization is not supported");
     }
 
-    private static StoringConverter primitiveConverter(PrimitiveType colSchema, boolean accum, boolean allowsNulls) {
+    private static StoringConverter primitiveConverter(
+            PrimitiveType colSchema,
+            boolean accum,
+            boolean dictionarySupport,
+            boolean allowsNulls) {
 
         PrimitiveType.PrimitiveTypeName name = colSchema.getPrimitiveTypeName();
         LogicalTypeAnnotation lt = colSchema.getLogicalTypeAnnotation();
@@ -57,11 +61,11 @@ public interface StoringConverter {
         if (lt != null) {
 
             if (lt.equals(LogicalTypeAnnotation.stringType())) {
-                return StringConverter.of(accum, DEFAULT_CAPACITY, allowsNulls);
+                return StringConverter.of(accum, DEFAULT_CAPACITY, dictionarySupport, allowsNulls);
             }
 
             if (lt.equals(LogicalTypeAnnotation.enumType())) {
-                return StringConverter.of(accum, DEFAULT_CAPACITY, allowsNulls);
+                return StringConverter.of(accum, DEFAULT_CAPACITY, dictionarySupport, allowsNulls);
             }
 
             if (lt instanceof LogicalTypeAnnotation.IntLogicalTypeAnnotation intType) {
@@ -71,8 +75,8 @@ public interface StoringConverter {
                     case 8 -> ByteConverter.of(accum, DEFAULT_CAPACITY, allowsNulls);
                     case 16 -> ShortConverter.of(accum, DEFAULT_CAPACITY, allowsNulls);
 
-                    case 32 -> IntConverter.of(accum, DEFAULT_CAPACITY, allowsNulls);
-                    case 64 -> LongConverter.of(accum, DEFAULT_CAPACITY, allowsNulls);
+                    case 32 -> IntConverter.of(accum, DEFAULT_CAPACITY, dictionarySupport, allowsNulls);
+                    case 64 -> LongConverter.of(accum, DEFAULT_CAPACITY, dictionarySupport, allowsNulls);
 
                     default -> throw new IllegalArgumentException("Invalid bit width for an int type: " + colSchema);
                 };
@@ -83,7 +87,7 @@ public interface StoringConverter {
                     throw new IllegalArgumentException("Can't decode as UUID, must be a 'FIXED_LEN_BYTE_ARRAY': " + colSchema);
                 }
 
-                return UuidConverter.of(accum, DEFAULT_CAPACITY, allowsNulls);
+                return UuidConverter.of(accum, DEFAULT_CAPACITY, dictionarySupport, allowsNulls);
             }
 
             if (lt.equals(LogicalTypeAnnotation.dateType())) {
@@ -91,14 +95,14 @@ public interface StoringConverter {
                     throw new IllegalArgumentException("Can't decode as DATE: " + colSchema);
                 }
 
-                return LocalDateConverter.of(accum, DEFAULT_CAPACITY, allowsNulls);
+                return LocalDateConverter.of(accum, DEFAULT_CAPACITY, dictionarySupport, allowsNulls);
             }
 
             if (lt instanceof LogicalTypeAnnotation.TimeLogicalTypeAnnotation time) {
                 return switch (name) {
-                    case INT32 -> LocalTimeMillisConverter.of(accum, DEFAULT_CAPACITY, allowsNulls);
+                    case INT32 -> LocalTimeMillisConverter.of(accum, DEFAULT_CAPACITY, dictionarySupport, allowsNulls);
                     case INT64 ->
-                            LocalTimeMicrosNanosConverter.of(accum, DEFAULT_CAPACITY, allowsNulls, time.getUnit());
+                            LocalTimeMicrosNanosConverter.of(accum, DEFAULT_CAPACITY, dictionarySupport, allowsNulls, time.getUnit());
                     default -> throw new IllegalArgumentException("Can't decode as TIME: " + colSchema);
                 };
             }
@@ -109,12 +113,12 @@ public interface StoringConverter {
                 }
 
                 return ts.isAdjustedToUTC()
-                        ? InstantConverter.of(accum, DEFAULT_CAPACITY, allowsNulls, ts.getUnit())
-                        : LocalDateTimeConverter.of(accum, DEFAULT_CAPACITY, allowsNulls, ts.getUnit());
+                        ? InstantConverter.of(accum, DEFAULT_CAPACITY, dictionarySupport, allowsNulls, ts.getUnit())
+                        : LocalDateTimeConverter.of(accum, DEFAULT_CAPACITY, dictionarySupport, allowsNulls, ts.getUnit());
             }
 
             if (lt instanceof LogicalTypeAnnotation.DecimalLogicalTypeAnnotation dt) {
-                return DecimalConverter.of(accum, DEFAULT_CAPACITY, allowsNulls, name, dt.getScale());
+                return DecimalConverter.of(accum, DEFAULT_CAPACITY, dictionarySupport, allowsNulls, name, dt.getScale());
             }
 
             if (lt instanceof LogicalTypeAnnotation.Float16LogicalTypeAnnotation) {
@@ -129,10 +133,10 @@ public interface StoringConverter {
         }
 
         return switch (name) {
-            case INT32 -> IntConverter.of(accum, DEFAULT_CAPACITY, allowsNulls);
-            case INT64 -> LongConverter.of(accum, DEFAULT_CAPACITY, allowsNulls);
-            case FLOAT -> FloatConverter.of(accum, DEFAULT_CAPACITY, allowsNulls);
-            case DOUBLE -> DoubleConverter.of(accum, DEFAULT_CAPACITY, allowsNulls);
+            case INT32 -> IntConverter.of(accum, DEFAULT_CAPACITY, dictionarySupport, allowsNulls);
+            case INT64 -> LongConverter.of(accum, DEFAULT_CAPACITY, dictionarySupport, allowsNulls);
+            case FLOAT -> FloatConverter.of(accum, DEFAULT_CAPACITY, dictionarySupport, allowsNulls);
+            case DOUBLE -> DoubleConverter.of(accum, DEFAULT_CAPACITY, dictionarySupport, allowsNulls);
             case BOOLEAN -> BoolConverter.of(accum, DEFAULT_CAPACITY, allowsNulls);
             case BINARY, FIXED_LEN_BYTE_ARRAY -> BytesConverter.of(accum, DEFAULT_CAPACITY, allowsNulls);
             case INT96 -> throw new RuntimeException("INT96 deserialization is deprecated and is not supported");